Move function contents
