+++ esphome/core/time.h
@@ -6,6 +6,12 @@
 
 namespace esphome {
 
+template<typename T> bool increment_time_value(T &current, uint16_t begin, uint16_t end);
+
+static bool is_leap_year(uint32_t year);
+
+static uint8_t days_in_month(uint8_t month, uint16_t year);
+
 /// A more user-friendly version of struct tm from time.h
 struct ESPTime {
   /** seconds after the minute [0-60]
